move delete facility logic from home to /facility
