+++ src/page-multi-facility/FacilityRow.tsx
@@ -6,9 +6,7 @@ import { saveFacility } from "../database/index";
 import Colors, { MarkColors as markColors } from "../design-system/Colors";
 import { DateMMMMdyyyy } from "../design-system/DateFormats";
 import iconEditSrc from "../design-system/icons/ic_edit.svg";
-import { StyledButton } from "../design-system/InputButton";
 import InputTextArea from "../design-system/InputTextArea";
-import ModalDialog from "../design-system/ModalDialog";
 import CurveChartContainer from "../impact-dashboard/CurveChartContainer";
 import {
   totalConfirmedCases,
@@ -23,42 +21,6 @@ const groupStatus = {
   hospitalized: true,
   infectious: true,
 };
-
-const ModalContents = styled.div`
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  font-weight: normal;
-  justify-content: flex-start;
-  margin-top: 30px;
-`;
-
-const ModalText = styled.div`
-  font-size: 13px;
-  margin-right: 25px;
-`;
-
-const ModalButtons = styled.div`
-  /* display: flex;
-  flex-direction: column; */
-`;
-
-const ModalButton = styled(StyledButton)`
-  font-size: 14px;
-  font-weight: normal;
-`;
-
-const DeleteButton = styled(ModalButton)`
-  background: ${Colors.darkRed};
-  color: ${Colors.white};
-  margin-right: 15px;
-`;
-
-const CancelButton = styled(ModalButton)`
-  background: transparent;
-  border: 1px solid ${Colors.forest};
-  color: ${Colors.forest};
-`;
 
 const FacilityNameLabel = styled.label`
   cursor: pointer;
@@ -101,13 +63,11 @@ const handleSubClick = (fn?: Function, ...args: any[]) => {
 };
 
 interface Props {
-  deleteFn: (scenarioId: string, facilityId: string) => void;
   facility: Facility;
   scenarioId: string;
 }
 
 const FacilityRow: React.FC<Props> = ({
-  deleteFn,
   facility: initialFacility,
   scenarioId: scenarioId,
 }) => {
@@ -118,21 +78,10 @@ const FacilityRow: React.FC<Props> = ({
 
   const { id, name, updatedAt } = facility;
 
-  const [showDeleteModal, updateShowDeleteModal] = useState(false);
-
   const openFacilityPage = () => {
     setFacility(facility);
     history.push("/facility");
   };
-
-  const openDeleteModal = handleSubClick(updateShowDeleteModal, true);
-
-  const closeDeleteModal = handleSubClick(updateShowDeleteModal, false);
-
-  const removeFacility = handleSubClick(async () => {
-    await deleteFn(scenarioId, id);
-    updateShowDeleteModal(false);
-  });
 
   return (
     <div onClick={openFacilityPage} className="cursor-pointer">
@@ -162,35 +111,11 @@ const FacilityRow: React.FC<Props> = ({
               <IconEdit alt="Edit facility name" src={iconEditSrc} />
             </FacilityNameLabel>
           </div>
-          <div className="text-xs text-gray-500 pb-4 flex flex-row justify-between">
+          <div className="text-xs text-gray-500 pb-4">
             <div>
               Last Update: <DateMMMMdyyyy date={new Date(updatedAt.toDate())} />
             </div>
-            <div className="mr-8">
-              <a className="px-1" href="#" onClick={openDeleteModal}>
-                Delete
-              </a>
-              <ModalDialog
-                closeModal={closeDeleteModal}
-                open={showDeleteModal}
-                title="Are you sure?"
-              >
-                <ModalContents>
-                  <ModalText>This action cannot be undone.</ModalText>
-                  <ModalButtons>
-                    <DeleteButton
-                      label="Delete facility"
-                      onClick={removeFacility}
-                    >
-                      Delete facility
-                    </DeleteButton>
-                    <CancelButton onClick={closeDeleteModal}>
-                      Cancel
-                    </CancelButton>
-                  </ModalButtons>
-                </ModalContents>
-              </ModalDialog>
-            </div>
+            <div className="mr-8" />
           </div>
         </div>
         <div className="w-3/5">
